Move RazorRenderer to SharpRazor and other rendering improvements
SharpRazor is a much better RazorEngine than RazorEngine. So we'll
move to that. While doing that, made several small improvements to all
the renders.

+++ test/IntegrationFixture.cs
@@ -84,12 +84,9 @@ namespace RobMensching.TinySite.Test
 
         private static void SafeDeleteFolder(string outputPath)
         {
-            try
+            if (Directory.Exists(outputPath))
             {
                 Directory.Delete(outputPath, true);
-            }
-            catch (DirectoryNotFoundException)
-            {
             }
         }
